requirements.txt
python-telegram-bot
groq
requests
pandas
telethon